requirements.txt
setuptools>=80.9.0
langchain>=0.1.0
langchain-community>=0.0.10
langchain-ollama>=0.0.1
pydantic>=2.0.0
pyyaml>=6.0.1
requests>=2.31.0
fastapi>=0.100.0
uvicorn>=0.23.0 